+++ src/VirtoCommerce.AuthorizeNet.Web/VirtoCommerce.AuthorizeNet.Web.csproj
@@ -1,4 +1,4 @@
-﻿ <Project Sdk="Microsoft.NET.Sdk.Web">
+ <Project Sdk="Microsoft.NET.Sdk.Web">
  	<PropertyGroup>
  		<TargetFramework>netcoreapp3.1</TargetFramework>
  		<ApplicationIcon />
